+++ lib/src/in_app_webview/main.dart
@@ -3,6 +3,5 @@ export 'in_app_webview.dart';
 export 'in_app_webview_controller.dart';
 export 'in_app_webview_options.dart';
 export 'headless_in_app_webview.dart';
-export '../pull_to_refresh/pull_to_refresh_controller.dart';
 export 'android/main.dart';
 export 'ios/main.dart';